% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{add_baseline_mean}
\alias{add_baseline_mean}
\title{Add Baseline Mean to Theta}
\usage{
add_baseline_mean(theta, M)
}
\arguments{
\item{theta}{Latent array}

\item{M}{Baseline mean (may be NULL)}
}
\value{
theta + M with appropriate broadcasting
}
\description{
Broadcasts M across time dimension of theta
}
\keyword{internal}
