% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{amprod}
\alias{amprod}
\title{Array-Matrix Product}
\usage{
amprod(A, M, k)
}
\arguments{
\item{A}{Array}

\item{M}{Matrix}

\item{k}{Mode}
}
\value{
Transformed array
}
\description{
Multiplies an array by a matrix along a given mode
}
\keyword{internal}
