% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{bilinear_step}
\alias{bilinear_step}
\title{Bilinear Multiplication Step}
\usage{
bilinear_step(A, Theta_prev, B)
}
\arguments{
\item{A}{Left matrix}

\item{Theta_prev}{Middle matrix}

\item{B}{Right matrix}
}
\value{
Bilinear product
}
\description{
Computes A \%\emph{\% Theta \%}\% t(B)
}
\keyword{internal}
