% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_z_update.R
\name{compute_gaussian_obs_residuals}
\alias{compute_gaussian_obs_residuals}
\title{Compute Gaussian Observation Residuals}
\usage{
compute_gaussian_obs_residuals(Z, Theta, M)
}
\arguments{
\item{Z}{Observation array (for Gaussian, Z = Y)}

\item{Theta}{Latent mean array}

\item{M}{Baseline mean array}
}
\value{
Sum of squared residuals
}
\description{
Sum of squared residuals for sigma2_obs update
}
\keyword{internal}
