% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_z_update.R
\name{compute_gaussian_obs_residuals_flat}
\alias{compute_gaussian_obs_residuals_flat}
\title{Vectorized Gaussian Residual Computation}
\usage{
compute_gaussian_obs_residuals_flat(Z_flat, Theta_flat, M_flat, m, p, Tt)
}
\arguments{
\item{Z_flat}{Flattened observation array}

\item{Theta_flat}{Flattened theta array}

\item{M_flat}{Flattened M array}

\item{m}{Number of nodes}

\item{p}{Number of relations}

\item{Tt}{Number of time points}
}
\value{
Sum of squared residuals
}
\description{
Vectorized version for flattened arrays
}
\keyword{internal}
