% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blocks.R
\name{compute_waic_dynamic}
\alias{compute_waic_dynamic}
\title{Compute WAIC for Dynamic Model}
\usage{
compute_waic_dynamic(fit)
}
\arguments{
\item{fit}{dynamic dbn fit}
}
\value{
list with waic, lppd, p_waic
}
\description{
WAIC computation for dynamic DBN
}
\keyword{internal}
