% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blocks.R
\name{compute_waic_piecewise}
\alias{compute_waic_piecewise}
\title{Compute WAIC for Piecewise Model}
\usage{
compute_waic_piecewise(fit)
}
\arguments{
\item{fit}{piecewise dbn fit}
}
\value{
list with waic, lppd, p_waic
}
\description{
WAIC computation for piecewise DBN
}
\keyword{internal}
