% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blocks.R
\name{compute_waic_static}
\alias{compute_waic_static}
\title{Compute WAIC for Static Model}
\usage{
compute_waic_static(fit)
}
\arguments{
\item{fit}{static dbn fit}
}
\value{
list with waic, lppd, p_waic
}
\description{
WAIC computation for static DBN
}
\keyword{internal}
