% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_core.R
\name{create_model_funs}
\alias{create_model_funs}
\title{Create Model Functions List}
\usage{
create_model_funs(model)
}
\arguments{
\item{model}{Model type}
}
\value{
List of update functions
}
\description{
Creates list of model-specific update functions
}
\keyword{internal}
