% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_structure.R, R/dag_structure_unexported.R
\name{dag_structure2}
\alias{dag_structure}
\alias{dag_structure2}
\alias{dag_structure_get_future_dependency}
\alias{dag_structure_get_node_str}
\alias{dag_structure_get_parent_list}
\title{Convert a formula into an adjacency matrix}
\usage{
dag_structure(fm, coll = NULL)

dag_structure_get_node_str(fm)

dag_structure_get_future_dependency(node_str)

dag_structure_get_parent_list(edge)
}
\arguments{
\item{fm}{A formula object defining the network}

\item{edge}{A character matrix. The first column is a node. The 
second column is character string of parents.}
}

