% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multichain.R
\name{dbn_multichain-methods}
\alias{dbn_multichain-methods}
\alias{coef.dbn_multichain}
\alias{predict.dbn_multichain}
\alias{gof.dbn_multichain}
\title{Methods for \code{dbn_multichain} objects}
\usage{
\method{coef}{dbn_multichain}(object, what = c("all", "A", "B", "M"), ...)

\method{predict}{dbn_multichain}(object, ...)

\method{gof}{dbn_multichain}(fit, ...)
}
\arguments{
\item{object, fit}{A \code{dbn_multichain} object.}

\item{what}{For \code{coef}, which operator slot to return.}

\item{...}{Forwarded to the underlying single-chain method.}
}
\value{
Whatever the underlying single-chain method returns.
}
\description{
S3 methods that combine the chains (via \code{\link[=combine_chains]{combine_chains()}}) and
delegate to the corresponding \code{dbn} method.
}
\author{
Tosin Salau and Shahryar Minhas
}
