% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{dyad_path}
\alias{dyad_path}
\title{Plot Dyad Trajectory}
\usage{
dyad_path(fit, i, j, rel = NULL, facet = TRUE, cred = c(0.025, 0.975))
}
\arguments{
\item{fit}{Dynamic dbn object}

\item{i}{Actor i index}

\item{j}{Actor j index}

\item{rel}{Relation indices (default: NULL = all relations)}

\item{facet}{Whether to facet by relation (default: TRUE)}

\item{cred}{Credible interval quantiles (default: c(0.025, 0.975))}
}
\value{
A ggplot2 object
}
\description{
Plots posterior mean and 95\% bands for a single dyad through time
}
\examples{
\donttest{
sim <- simulate_dynamic_dbn(n = 6, time = 10, seed = 1)
fit <- dbn(sim$Y, model = "dynamic", nscan = 200, burn = 100, verbose = FALSE)
dyad_path(fit, i = 1, j = 2)
}
}
\seealso{
\code{\link{dbn}}, \code{\link{net_snapshot}},
\code{\link{role_trajectory}}, \code{\link{theta_summary}}
}
\author{
Tosin Salau and Shahryar Minhas
}
