% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_memory}
\alias{estimate_memory}
\title{Estimate Memory Usage for Dynamic DBN}
\usage{
estimate_memory(
  n_row,
  n_col = n_row,
  p = 1,
  Tt = 50,
  nscan = 5000,
  burn = 1000,
  odens = 1,
  time_thin = 1,
  family = "ordinal",
  quiet = FALSE
)
}
\arguments{
\item{n_row}{Number of sender nodes}

\item{n_col}{Number of receiver nodes (default: n_row)}

\item{p}{Number of relations}

\item{Tt}{Number of time points}

\item{nscan}{Number of MCMC iterations after burn-in}

\item{burn}{Number of burn-in iterations}

\item{odens}{Thinning interval}

\item{time_thin}{Time-thinning interval}

\item{family}{Data family ("ordinal", "gaussian", "binary")}

\item{quiet}{Suppress printed output}
}
\value{
Estimated memory in GB (invisibly)
}
\description{
Estimates RAM (in GB) needed for dynamic model output storage
}
\examples{
# Estimate memory for a moderate network
estimate_memory(n_row = 20, Tt = 30, nscan = 5000)

# Bipartite network with time thinning
estimate_memory(n_row = 15, n_col = 25, Tt = 50,
                nscan = 10000, time_thin = 5)

# Suppress printed output, just get the value
gb <- estimate_memory(n_row = 10, Tt = 20, nscan = 2000, quiet = TRUE)
gb
}
\seealso{
\code{\link{dbn}}, \code{\link{dbn_dynamic}}
}
