% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_node.R
\name{expand_node}
\alias{expand_node}
\title{Convert Root Node to Temporal Node}
\usage{
expand_node(node, time, dynamic)
}
\arguments{
\item{node}{\code{character(1)} naming a root node}

\item{time}{\code{numeric(1)} giving \code{max_t} value for the node}

\item{dynamic}{\code{logical(1)} indicating if the node is a dynamic node}
}
\description{
Appends the \code{_n} suffix to the root node. Presumably,
  this is only of interest in plotting the expanded network.
}
\section{Functional Requirements}{

\enumerate{
  \item If \code{dynamic = TRUE}, returns a character vector where
    each element follows the format \code{"node_t"}.
  \item If \code{dynamic = FALSE}, returns \code{node}.
}
}
\author{
Benjamin Nutter
}

