% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{extract_baseline}
\alias{extract_baseline}
\title{Extract Baseline Network for IRF}
\usage{
extract_baseline(fit, draw_idx, n_row, n_col, p)
}
\arguments{
\item{fit}{dbn model fit object}

\item{draw_idx}{Posterior draw index}

\item{n_row}{Number of sender nodes}

\item{n_col}{Number of receiver nodes}

\item{p}{Number of relations}
}
\value{
Baseline network matrix
}
\description{
Extract Baseline Network for IRF
}
\keyword{internal}
