% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family_binary}
\alias{family_binary}
\title{Binary Family}
\usage{
family_binary()
}
\value{
A dbn_family object
}
\description{
Probit-link family for binary outcomes.
Observation variance fixed at 1 for identifiability.
}
\keyword{internal}
