% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ffbs.R
\name{ffbs_dlm}
\alias{ffbs_dlm}
\title{Forward-Filter Backward-Sample for DLM}
\usage{
ffbs_dlm(y, Flist, V, W, m0, C0, ar1 = FALSE, rho = 0)
}
\arguments{
\item{y}{List of observation vectors}

\item{Flist}{List of design matrices}

\item{V}{Observation variance matrix}

\item{W}{State innovation variance matrix}

\item{m0}{Prior mean vector}

\item{C0}{Prior covariance matrix}

\item{ar1}{Logical: use AR(1) dynamics}

\item{rho}{AR(1) coefficient (used if ar1=TRUE)}
}
\value{
Matrix of sampled state vectors (columns are time points)
}
\description{
FFBS for dynamic linear model with optional AR(1) dynamics
}
\keyword{internal}
