% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ffbs.R
\name{ffbs_theta}
\alias{ffbs_theta}
\title{FFBS for Theta Matrix}
\usage{
ffbs_theta(Z, mu, Aarray, Barray, sigma2)
}
\arguments{
\item{Z}{Observed data (m x m x Tt)}

\item{mu}{Baseline mean (m x m)}

\item{Aarray}{Time-varying A matrices (m x m x Tt)}

\item{Barray}{Time-varying B matrices (m x m x Tt)}

\item{sigma2}{Innovation variance}
}
\value{
Array of sampled Theta matrices (m x m x Tt)
}
\description{
FFBS for the latent Theta process in bilinear model
}
\keyword{internal}
