% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynamic_time.R
\name{get_dynamic_time}
\alias{get_dynamic_time}
\title{Get the Time Suffix from a Dynamic Node}
\usage{
get_dynamic_time(node)
}
\arguments{
\item{node}{\code{character(1)} giving a node name.}
}
\description{
Returns the integer suffix appended to a dynamic node.
}
\section{Functional Requirements}{

\enumerate{
  \item Return the integer suffix at the end of a node name.
  \item If the node is not dynamic, return \code{numeric(0)}.
}
}
\author{
Benjamin Nutter
}

