% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{get_family}
\alias{get_family}
\title{Get family object from model fit}
\usage{
get_family(fit)
}
\arguments{
\item{fit}{A dbn model fit object}
}
\value{
A dbn_family object or NULL
}
\description{
Extract or reconstruct family object from model fit
}
\keyword{internal}
