% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blocks.R
\name{get_min_block_length}
\alias{get_min_block_length}
\title{Get Minimum Block Length}
\usage{
get_min_block_length(n)
}
\arguments{
\item{n}{number of actors}
}
\value{
minimum recommended block length
}
\description{
determines minimum block length based on network size
}
\keyword{internal}
