% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_formula.R
\name{get_node_formula}
\alias{get_node_formula}
\title{Retrieve the node definition from the node formula}
\usage{
get_node_formula(network, node = NULL)
}
\arguments{
\item{network}{An object of class \code{dbn}.}

\item{node}{A character vector of nodes for which the formula definitions 
are to be obtained.}
}
\description{
Retrieve the node definition from a \code{dbn} object node.
}
\section{Functional Requirements}{

\enumerate{
  \item Return the raw node definition from the \code{node_attr} data
      frame for each node listed in \code{node}.
  \item If any \code{node} is not a node in \code{network}, cast a warning
      and drop the invalid nodes.
  \item If \code{node} is NULL, return all of the node formulae.
  \item If \code{network} is not a \code{dbn}, cast an error.
}
}

