% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_utilties.R, R/is_node_utilities.R
\name{get_node_utilities}
\alias{get_node_decision}
\alias{get_node_deterministic}
\alias{get_node_dynamic}
\alias{get_node_maxt}
\alias{get_node_parent}
\alias{get_node_utilities}
\alias{get_node_utility}
\title{Utilities to Get Node Attributes}
\usage{
get_node_parent(network, node = NULL)

get_node_maxt(network, node = NULL)

get_node_decision(network, node = NULL)

get_node_deterministic(network, node = NULL)

get_node_dynamic(network, node = NULL)

get_node_utility(network, node = NULL)
}
\arguments{
\item{network}{An object of class \code{dbn}.}

\item{node}{A character vector of nodes for which the formula definitions 
are to be obtained. If \code{NULL}, a logical for each node is returned.}
}
\description{
Methods to quickly get attributes
  of a node.  These may all be done through subsetting the 
  \code{node_attr} attribute of the network object.  These functions
  perform the work in a standardized way, allowing less typing.
}
\details{
\code{get_node_parent} returns a list giving the parents of the 
  requested nodes.
  
\code{get_node_dynamic} Get the dynamic attribute of the node.
  
\code{is_node_decision} Get the decision attribute of the node.

\code{is_node_utility} Get the utility attribute of the node.

\code{is_node_deterministic} Get the deterministic attribute of the node.
}
\section{Functional Requirements}{

\emph{get_node_parent}

\enumerate{
 \item Return a list of character vectors naming the parents of 
   each node in \code{node}.
 \item If \code{node = NULL}, \code{node} is assumed to be the 
    vector of all node names.
  \item If any \code{node} is not a node in \code{network}, cast a warning
      and drop the invalid nodes.
  \item If \code{network} is not a \code{dbn}, cast an error.  
}

\emph{get_node_maxt}

\enumerate{
 \item Return a numeric vector giving the value of \code{max_t} for 
    each node in \code{node}.
 \item If \code{node = NULL}, \code{node} is assumed to be the 
    vector of all node names.
  \item If any \code{node} is not a node in \code{network}, cast a warning
      and drop the invalid nodes.
  \item If \code{network} is not a \code{dbn}, cast an error.  
}

Requirements for \code{get_node_dynamic}, \code{get_node_decision},
\code{get_node_utility} and \code{get_node_deterministic} are given 
in \code{\link{is_node_utilities}}.  The \code{get} functions are 
aliases of the \code{is} functions.
}
\seealso{
\code{\link{set_node_utilities}}, \code{\link{is_node_utilities}}
}

