% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Default value if a is NULL}
}
\value{
a if not NULL, otherwise b
}
\description{
Null Coalescing Operator
}
\keyword{internal}
