% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{\%sp\%}
\alias{\%sp\%}
\title{Sparse Matrix Multiply Operator}
\usage{
A \%sp\% B
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}
}
\value{
Matrix product
}
\description{
Infix operator that keeps sparse objects sparse
}
\keyword{internal}
