% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impulse_response_const}
\alias{impulse_response_const}
\title{Compute impulse response for constant A,B matrices}
\usage{
impulse_response_const(A, B, S, H)
}
\arguments{
\item{A}{Transition matrix A (m x m)}

\item{B}{Transition matrix B (m x m)}

\item{S}{Shock matrix (m x m)}

\item{H}{Number of horizons to compute}
}
\value{
Cube of impulse responses (m x m x H+1)
}
\description{
Compute impulse response for constant A,B matrices
}
