% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impulse_response_dynamic}
\alias{impulse_response_dynamic}
\title{Compute impulse response for time-varying A,B matrices}
\usage{
impulse_response_dynamic(Aarray, Barray, S, t0, H)
}
\arguments{
\item{Aarray}{Cube of A matrices over time (m x m x T)}

\item{Barray}{Cube of B matrices over time (m x m x T)}

\item{S}{Shock matrix (m x m)}

\item{t0}{Time index of shock (0-based)}

\item{H}{Number of horizons to compute}
}
\value{
Cube of impulse responses (m x m x H+1)
}
\description{
Compute impulse response for time-varying A,B matrices
}
