% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_lowrank.R
\name{init_lowrank}
\alias{init_lowrank}
\title{Initialize low-rank model}
\usage{
init_lowrank(Y, r)
}
\arguments{
\item{Y}{Data array}

\item{r}{Rank}
}
\value{
List with U and alpha
}
\description{
Initialize low-rank model
}
\keyword{internal}
