% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_node_something.R
\name{is_node_something}
\alias{get_node_decision}
\alias{get_node_deterministic}
\alias{get_node_dynamic}
\alias{get_node_parent}
\alias{get_node_utility}
\alias{is_node_decision}
\alias{is_node_deterministic}
\alias{is_node_dynamic}
\alias{is_node_parent}
\alias{is_node_parent_of}
\alias{is_node_something}
\alias{is_node_utility}
\title{Utilities to identify node attributes}
\usage{
is_node_parent(network, node = NULL)

get_node_parent(network, node = NULL)

is_node_parent_of(network, node = NULL, child)

is_node_dynamic(network, node = NULL)

get_node_dynamic(network, node = NULL)

is_node_decision(network, node = NULL)

get_node_decision(network, node = NULL)

is_node_utility(network, node = NULL)

get_node_utility(network, node = NULL)

is_node_deterministic(network, node = NULL)

get_node_deterministic(network, node = NULL)
}
\arguments{
\item{network}{An object of class \code{dbn}.}

\item{node}{A character vector of nodes for which the formula definitions 
are to be obtained. If \code{NULL}, a logical for each node is returned.}

\item{child}{A \code{character(1)} vector used in \code{is_node_parent_of}. 
This is the node for which \code{node} are being tested as parents.}
}
\description{
Methods to quickly get logical indications of the attributes
  of a node.  These may all be done through subsetting the 
  \code{node_attr} attribute of the network object.  These functions
  perform the work in a standardized way, allowing less typing.
}
\details{
\code{is_node_parent} returns a logical stating if the node is a parent of
  at least one other node.
  
\code{get_node_parent} returns a list of character vectors. Each element
  of the list is the parents for the corresponding value of \code{node}.
  
\code{is_node_parent_of} returns a logical stating if node is a parent of 
  at least one element of \code{child}
  
\code{is_node_dynamic} Is the node a dynamic node. \code{get_node_dynamic}
is an alias.
  
\code{is_node_decision} Is the node a decision node. \code{get_node_decision}
is an alias.

\code{is_node_utility} Is the node a utility node. \code{get_node_utility}
is an alias.

\code{is_node_deterministic} Is the node a deterministic node.  
\code{get_node_deterministic} is an alias.

\code{is_node_something} is an unexported generalized form from which 
\code{is_node_dynamic}, \code{is_node_decision}, \code{is_node_utility},
\code{is_node_deterministic} are derived.  This is done to minimize the
code base that needs to be maintained.
}
\section{Functional Requirements}{

\enumerate{
  \item Each method returns a logical vector (except for \code{get_node_parent}).
  \item If \code{node = NULL}, \code{node} is assumed to be the 
    vector of all node names.
  \item If any \code{node} is not a node in \code{network}, cast a warning
      and drop the invalid nodes.
  \item If \code{network} is not a \code{dbn}, cast an error.
  \item For \code{is_node_parent_of}, if \code{child} has length not 
      equal to 1, cast an error.
  \item For \code{get_node_parent}, returns a list of character vectors.
}
}

