% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_syntactic.R
\name{is_syntactic}
\alias{is_syntactic}
\title{Check for Syntactically Valid R Names}
\source{
\url{http://r.789695.n4.nabble.com/Syntactically-valid-names-td3636819.html}
}
\usage{
is_syntactic(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A logical vector with the same length as \code{x}
}
\description{
Determines if the elements of a character string are 
  syntactically valid names.
}
\section{Functional Requirements}{

\enumerate{
  \item Return a logical vector the same length as \code{x}
  \item Cast an error if \code{x} is not a character vector
}
}
\author{
Hadley Wickham
}

