% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-Sum-Exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
log(sum(exp(x)))
}
\description{
Numerically stable log-sum-exp computation
}
\keyword{internal}
