% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{mat}
\alias{mat}
\title{Matrix Unfold Operation}
\usage{
mat(X, mode)
}
\arguments{
\item{X}{Array to unfold}

\item{mode}{Mode along which to unfold (1, 2, or 3)}
}
\value{
Matrix with mode as rows
}
\description{
Unfolds an array along a specified mode
}
\keyword{internal}
