% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_form_to_node_form.R
\name{model_form_to_node_form}
\alias{model_form_to_node_form}
\alias{model_form_to_node_form.default}
\alias{model_form_to_node_form.xtabs}
\title{Convert a Model Object Formula to a Node Formula}
\usage{
model_form_to_node_form(fit, ...)

\method{model_form_to_node_form}{default}(fit, ...)

\method{model_form_to_node_form}{xtabs}(fit, ...)
}
\arguments{
\item{fit}{An object with a \code{call} element}

\item{...}{Arguments to pass to other methods.  Currently ignored.}
}
\value{
Returns a character string with the converted formula
}
\description{
A utility for \code{dbn.list} to take the model
  formula of an object and convert it to a format that can be used 
  in creating a \code{dbn} object. Different objects may require
  different methods to get the formula, so these are written as 
  methods.
}
\section{Functional Requirements}{

\enumerate{
  \item Returns a character string with a converted formula
  \item Require that \code{xtabs} models only have one variable.
  \item Should formula extraction fail, return an error message with
        directions to file an issue on GitHub.
}
}
\author{
Benjamin Nutter
}

