% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_node.R
\name{model_to_node}
\alias{model_to_node}
\alias{model_to_node.default}
\alias{model_to_node.xtabs}
\title{Convert a Model to a Node Description}
\usage{
model_to_node(m, ...)

\method{model_to_node}{default}(m, ...)

\method{model_to_node}{xtabs}(m, ...)
}
\arguments{
\item{m}{A model object}

\item{...}{Additional arguments to pass to other methods}
}
\description{
Write a node definition to be placed in a network formula. 
  This is used when creating networks from a list of models.
}
\section{Functional Requirements}{

Default Method
\enumerate{
  \item Cast an error if the model formula has more than one element on 
        the left hand side.
  \item Assume any node in a function is given as the first argument
        of the function unless the function is \code{Surv}.
}

Additional Requirement for the \code{xtabs} method
\enumerate{
  \item Only allow one variable in \code{xtabs} definitions 
}
}

