% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_node_unexported.R
\name{model_to_node_unexported}
\alias{model_to_node_deparse_side}
\alias{model_to_node_unexported}
\title{Unexported Utilities for \code{model_to_node}}
\usage{
model_to_node_deparse_side(side)
}
\arguments{
\item{side}{A left or right hand side of a formula object.  The object 
should be deparsed and split on \code{~} before passing into 
\code{mode_to_node_deparse_side}.}
}
\description{
Support functions for \code{model_to_node}. These assist with
parsing the formula object from a model into a node definition to include 
in \code{dbn}.
}
\seealso{
\code{\link{model_to_node}}
}

