% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{plot.dbn_forecast_ci}
\alias{plot.dbn_forecast_ci}
\title{Plot method for dbn_forecast_ci}
\usage{
\method{plot}{dbn_forecast_ci}(x, max_cells = 16L, ...)
}
\arguments{
\item{x}{A \code{dbn_forecast_ci} list.}

\item{max_cells}{Maximum number of (sender, receiver) facets.}

\item{...}{Unused.}
}
\value{
Invisibly, the ggplot object.
}
\description{
Identical to \code{\link[=plot.dbn_forecast]{plot.dbn_forecast()}} but renders the lower / upper
bands as a ribbon around the mean.
}
\author{
Tosin Salau and Shahryar Minhas
}
