% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{plot.dbn_role_trajectory}
\alias{plot.dbn_role_trajectory}
\title{Plot method for \code{dbn_role_trajectory}}
\usage{
\method{plot}{dbn_role_trajectory}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_role_trajectory} object from \code{\link[=role_trajectory]{role_trajectory()}}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot method for \code{dbn_role_trajectory}
}
\author{
Tosin Salau and Shahryar Minhas
}
