% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dbn.R
\name{plot.dbn_unexported}
\alias{expand_dynamic}
\alias{get_edge_text}
\alias{get_node_string}
\alias{get_rank_string}
\alias{get_shared_parent}
\alias{get_star_node}
\alias{plot.dbn_unexported}
\alias{sanitize_custom_edge}
\alias{write_edge}
\title{Unexported Utilities for \code{plot.dbn}}
\usage{
expand_dynamic(Node)

get_rank_string(Node)

get_node_string(node_name, custom_node)

get_star_node(node_name, custom_node)

sanitize_custom_edge(custom_edge, custom_edge_name, Node)

write_edge(Node, custom_edge = custom_edge)

get_shared_parent(node, parent, Node)

get_edge_text(parent, node, shared_parent, custom_edge)
}
\arguments{
\item{Node}{The node attibutes data frame.}

\item{node_name}{the \code{Node$node_name} vector.}

\item{custom_node}{The \code{custom_node} argument from \code{plot.dbn}}

\item{node}{\code{character(1)} giving a node in the network.}

\item{parent}{\code{character(1)} giving a parent of \code{node}}

\item{shared_parent}{The \code{tbl_df} object
returned by \code{get_shared_parent}}
}
\description{
Documentation for utilities that support the 
  \code{plot.dbn} method.
}
\details{
\code{plot.dbn} will call \code{write_edge}, which in turn calls
  \code{get_shared_parent} and \code{get_edge_text}.  The shared 
  parents of two nodes are needed in order to determine when the 
  \code{"[constraint=false]"} tag should be added to an edge.
  This tag helps keep the dynamic nodes in proper alignment.
}
\section{Functional Requirements}{

\enumerate{
  \item \code{write_edge} returns one character string for each edge 
    in the network.
  \item \code{get_shared_parent} returns a vector naming the parents shared
    between two nodes.
  \item \code{get_edge_text} returns a character string defining a single 
    edge.  
  \item \code{get_edge_text} When drawing an edge between two nodes, if 
    the parent is a shared parent, the \code{"[constraint=false]"} tag 
    is added to the edge.
}
}
\author{
Benjamin Nutter
}

