% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{plot_dbn}
\alias{plot_dbn}
\title{Plot method for DBN objects}
\usage{
plot_dbn(x, ...)
}
\arguments{
\item{x}{A dbn object}

\item{...}{Additional arguments passed to model-specific plot functions}
}
\value{
Invisibly returns the plot object
}
\description{
Wrapper function for plot.dbn method
}
