% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_methods.R
\name{plot_hmm}
\alias{plot_hmm}
\title{Plot HMM-DBN Results}
\usage{
plot_hmm(x)
}
\arguments{
\item{x}{A dbn object with model="hmm"}
}
\value{
List of ggplot objects or combined plot
}
\description{
Diagnostic plots for HMM-DBN model fits
}
\keyword{internal}
