% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{plot_piecewise_blocks}
\alias{plot_piecewise_blocks}
\title{Block Comparison Plot}
\usage{
plot_piecewise_blocks(x, ...)
}
\arguments{
\item{x}{piecewise dbn fit}

\item{...}{additional arguments}
}
\description{
visualizes A matrix differences across blocks
}
\keyword{internal}
