% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{plot_piecewise_influence}
\alias{plot_piecewise_influence}
\title{Plot Influence Matrix for Block}
\usage{
plot_piecewise_influence(x, block = 1, ...)
}
\arguments{
\item{x}{piecewise dbn fit}

\item{block}{block index or name}

\item{...}{unused}
}
\value{
A \code{ggplot} object.
}
\description{
Heatmap of the posterior mean A for a single block.
}
\keyword{internal}
