% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{plot_ppc_bars}
\alias{plot_ppc_bars}
\title{Posterior predictive bar plot for discrete data}
\usage{
plot_ppc_bars(fit, ppd, rel = 1, time = NULL, Y_obs = NULL)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{ppd}{Posterior predictive samples}

\item{rel}{Relation index}

\item{time}{Time indices}

\item{Y_obs}{Observed data array (required)}
}
\value{
A ggplot object or NULL
}
\description{
Bar plot comparison for discrete outcomes
}
\keyword{internal}
