% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{ppc_ecdf}
\alias{ppc_ecdf}
\title{Posterior Predictive ECDF Overlay}
\usage{
ppc_ecdf(fit, n_rep = 20)
}
\arguments{
\item{fit}{Object of class "dbn"}

\item{n_rep}{Number of replicated datasets to draw (default: 20)}
}
\value{
A ggplot2 object
}
\description{
Creates posterior predictive checks using ECDF overlays
}
\keyword{internal}
