% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ranklik.R
\name{precompute_ranks}
\alias{precompute_ranks}
\title{Precompute Rank Indices}
\usage{
precompute_ranks(Y)
}
\arguments{
\item{Y}{Ordinal data array}
}
\value{
List of index lists by relation
}
\description{
Precomputes indices for each rank level to speed up MCMC
}
\keyword{internal}
