% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{predict_ordinal}
\alias{predict_ordinal}
\title{Posterior Predictive Ordinal Data}
\usage{
predict_ordinal(fit, draws = 100, H = NULL)
}
\arguments{
\item{fit}{A "dbn" object}

\item{draws}{Number of draws}

\item{H}{Forecast horizon (dynamic models)}
}
\value{
Array of ordinal predictions
}
\description{
Generate ordinal data from posterior predictive distribution
}
\keyword{internal}
