% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dbn.R
\name{print.dbn}
\alias{print.dbn}
\title{Print a \code{dbn} Object}
\usage{
\method{print}{dbn}(x, ...)
}
\arguments{
\item{x}{A \code{dbn} object.}

\item{...}{Additional arguments to pass to other methods. Currently ignored.}
}
\description{
Prints the node formulae of a \code{dbn} network.
}

