% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{print.dbn_boot}
\alias{print.dbn_boot}
\title{Print Bootstrap Results for ALS}
\usage{
\method{print}{dbn_boot}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A \code{dbn_boot} object from \code{\link[=dbn_explore_bootstrap]{dbn_explore_bootstrap()}}.}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the \code{dbn_boot} object.
}
\description{
Print Bootstrap Results for ALS
}
