% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_covariates_accessors.R
\name{print.dbn_covariates_fit}
\alias{print.dbn_covariates_fit}
\title{Print method for dbn_covariates_fit}
\usage{
\method{print}{dbn_covariates_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_covariates_fit} object.}

\item{...}{Unused.}
}
\description{
Print method for dbn_covariates_fit
}
\author{
Tosin Salau and Shahryar Minhas
}
