% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_extract.R
\name{print.dbn_derived}
\alias{print.dbn_derived}
\title{Print method for derived quantities}
\usage{
\method{print}{dbn_derived}(x, ...)
}
\arguments{
\item{x}{An object of class "dbn_derived"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for derived quantities
}
\author{
Tosin Salau and Shahryar Minhas
}
