% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{print.dbn_fevd}
\alias{print.dbn_fevd}
\title{Print method for dbn_fevd objects}
\usage{
\method{print}{dbn_fevd}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_fevd} object.}

\item{...}{Unused.}
}
\description{
Print method for dbn_fevd objects
}
\author{
Tosin Salau and Shahryar Minhas
}
