% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{print.dbn_forecast}
\alias{print.dbn_forecast}
\title{Print a dbn forecast object}
\usage{
\method{print}{dbn_forecast}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_forecast} array, returned by \code{\link[=predict.dbn]{predict.dbn()}} / \code{\link[=forecast.dbn]{forecast.dbn()}} with \code{H} set.}

\item{...}{Unused.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print a dbn forecast object
}
\author{
Tosin Salau and Shahryar Minhas
}
