% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{print.dbn_forecast_ci}
\alias{print.dbn_forecast_ci}
\title{Print a dbn forecast object with intervals}
\usage{
\method{print}{dbn_forecast_ci}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_forecast_ci} list of mean / lower / upper arrays.}

\item{...}{Unused.}
}
\value{
Invisibly, \code{x}.
}
\description{
Print a dbn forecast object with intervals
}
\author{
Tosin Salau and Shahryar Minhas
}
