% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{print.dbn_gof}
\alias{print.dbn_gof}
\title{Print method for dbn_gof}
\usage{
\method{print}{dbn_gof}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_gof} object.}

\item{...}{Unused.}
}
\description{
Print method for dbn_gof
}
\keyword{internal}
