% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multichain.R
\name{print.dbn_multichain}
\alias{print.dbn_multichain}
\title{Print method for dbn_multichain}
\usage{
\method{print}{dbn_multichain}(x, ...)
}
\arguments{
\item{x}{A \code{dbn_multichain} object.}

\item{...}{Unused.}
}
\description{
Print method for dbn_multichain
}
\keyword{internal}
