% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr_and_ranks.R
\name{print.dbn_operator}
\alias{print.dbn_operator}
\title{Print method for the DBN time-varying operator}
\usage{
\method{print}{dbn_operator}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{dbn_operator} object.}

\item{digits}{Number of digits (default 3).}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for the DBN time-varying operator
}
\author{
Tosin Salau and Shahryar Minhas
}
