% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_hmm.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Sample Dirichlet}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{Parameter vector}
}
\value{
Sample vector
}
\description{
Samples from Dirichlet distribution
}
\keyword{internal}
