% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_core.R
\name{register_dbn_model}
\alias{register_dbn_model}
\title{Register DBN Model}
\usage{
register_dbn_model(name, fun_list)
}
\arguments{
\item{name}{Model name}

\item{fun_list}{List of model functions}
}
\description{
Register a new DBN model variant
}
\examples{
\dontrun{
custom_model <- list(
    init = function(data) list(theta = rnorm(10)),
    update_Z = function(state, data) state,
    update_Theta = function(state, data) {
        state$theta <- state$theta + rnorm(10, 0, 0.1)
        state
    },
    collect = function(state) list(theta = state$theta)
)
register_dbn_model("custom", custom_model)
}
}
\keyword{internal}
