% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{rsan}
\alias{rsan}
\title{Random Standard Array Normal}
\usage{
rsan(d)
}
\arguments{
\item{d}{Dimensions of array}
}
\value{
Array of standard normal values
}
\description{
Generates array of standard normal random variables
}
\keyword{internal}
