% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{safe_rinv_gamma}
\alias{safe_rinv_gamma}
\title{Safe Inverse-Gamma Sampling}
\usage{
safe_rinv_gamma(shape, rate, floor = 1e-08, ceiling = 1e+08)
}
\arguments{
\item{shape}{Shape parameter}

\item{rate}{Rate parameter}

\item{floor}{Minimum return value}

\item{ceiling}{Maximum return value}
}
\value{
Sample from inverse-gamma distribution
}
\description{
Draws from inverse-Gamma on log-scale to avoid under/overflow
}
\keyword{internal}
