% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_model.R
\name{set_node_model}
\alias{has_node_model}
\alias{set_node_model}
\title{Set and Retrieve Models for \code{dbn} Objects}
\usage{
has_node_model(network, node = NULL)
}
\arguments{
\item{network}{An object of class \code{dbn}}

\item{node}{A character vector naming nodes in the network.}
}
\description{
Access and replace models associated with nodes in a \code{dbn}
  network.  This provides a more convenient interface with which to 
  make updates to the network.
}
\section{Functional Requirements}{

\emph{has_node_model}

\enumerate{
  \item If \code{node = NULL} or assume all nodes in the network
  \item If any \code{node} is not a node in \code{network}, cast a warning
      and drop the invalid nodes.
  \item Return a logical vector the length of \code{node} where \code{TRUE}
    indicates if the node has a model in the network.
  \item Cast an error if \code{network} is not a \code{dbn} object.
  \item Cast an error if \code{node} is not a character vector.
}
}

