% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_core.R
\name{shared_preprocess}
\alias{shared_preprocess}
\title{Shared Preprocessing}
\usage{
shared_preprocess(Y, family = "ordinal")
}
\arguments{
\item{Y}{Data array}

\item{family}{Distribution family}
}
\value{
List with preprocessed components
}
\description{
Common preprocessing for all DBN models
}
\keyword{internal}
