% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{simulate_piecewise}
\alias{simulate_piecewise}
\title{Simulate from Piecewise Model}
\usage{
simulate_piecewise(
  object,
  H = 10,
  ndraws = 100,
  S = NULL,
  summary = c("mean", "none"),
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{piecewise dbn fit}

\item{H}{forecast horizon}

\item{ndraws}{number of posterior draws to use}

\item{seed}{random seed}

\item{...}{additional arguments}
}
\value{
list with simulated values
}
\description{
generates forecasts or simulations from piecewise model
}
\keyword{internal}
