% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Sparse-Aware Matrix Multiply}
\usage{
sparse_mult(A, B)
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}
}
\value{
Product A \%*\% B
}
\description{
Multiplies matrices that may be sparse
}
\keyword{internal}
