% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_static.R
\name{static_gibbs_step_tucker}
\alias{static_gibbs_step_tucker}
\title{Static Model MCMC Using Tucker Products}
\usage{
static_gibbs_step_tucker(Y_obs, Z, M, B, s2, t2, g2)
}
\arguments{
\item{Y_obs}{observed data array (n_row x n_col x p x Tt)}

\item{Z}{latent continuous array (n_row x n_col x p x Tt)}

\item{M}{baseline mean (n_row x n_col x p)}

\item{B}{list of K influence matrices}

\item{s2}{observation variance}

\item{t2}{prior precision for B}

\item{g2}{prior variance for M}
}
\value{
list with updated B, M, s2, t2, g2, Y_theta, X_lag
}
\description{
alternative static model implementation using full Tucker products
}
\keyword{internal}
