% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{summary_dynamic}
\alias{summary_dynamic}
\title{Summary for dynamic DBN fits}
\usage{
summary_dynamic(object, digits = 3, ...)
}
\arguments{
\item{object}{Object of class "dbn" with model="dynamic"}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisible object
}
\description{
Prints summary statistics for dynamic DBN model results
}
\keyword{internal}
