% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_methods.R
\name{summary_hmm}
\alias{summary_hmm}
\title{Summary for HMM-DBN Fits}
\usage{
summary_hmm(object, digits = 3, ...)
}
\arguments{
\item{object}{A dbn object with model="hmm"}

\item{digits}{Number of digits to display}

\item{...}{Ignored}
}
\value{
Invisibly returns the input object
}
\description{
Posterior summaries for HMM-DBN models
}
\keyword{internal}
