% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{summary_piecewise}
\alias{summary_piecewise}
\title{Summary for Piecewise DBN}
\usage{
summary_piecewise(object, ...)
}
\arguments{
\item{object}{piecewise dbn fit}

\item{...}{additional arguments}
}
\value{
summary object
}
\description{
generates summary statistics for piecewise model
}
\keyword{internal}
