% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{summary_static}
\alias{summary_static}
\title{Summary for static DBN fits}
\usage{
summary_static(object, digits = 3, ...)
}
\arguments{
\item{object}{Object of class "dbn" with model="static"}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisible object
}
\description{
Prints summary statistics for static DBN model results
}
\keyword{internal}
