% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{tprod}
\alias{tprod}
\title{Tensor-Matrix Product}
\usage{
tprod(A, B, modes = 1:length(B))
}
\arguments{
\item{A}{Input tensor}

\item{B}{List of matrices}

\item{modes}{Modes for multiplication}
}
\value{
Transformed tensor
}
\description{
Computes tensor product with matrices along specified modes
}
\keyword{internal}
