% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_static.R
\name{update_B_tucker}
\alias{update_B_tucker}
\title{Update B Matrices Using Tucker Products}
\usage{
update_B_tucker(Y, X, B, s2, t2, K = length(B))
}
\arguments{
\item{Y}{response array (n_row x n_col x p x Tt)}

\item{X}{predictor array (n_row x n_col x p x Tt)}

\item{B}{list of K influence matrices}

\item{s2}{observation variance}

\item{t2}{prior precision for B}

\item{K}{number of modes (default 3)}
}
\value{
list with updated B matrices and sse
}
\description{
updates all K influence matrices using Tucker product formulation
}
\keyword{internal}
