% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opt.R
\name{update_M_fast}
\alias{update_M_fast}
\title{Update M with Optimal Memory Usage}
\usage{
update_M_fast(Z, g2)
}
\arguments{
\item{Z}{Latent values array (n_row x n_col x p x Tt)}

\item{g2}{Prior variance for M}
}
\value{
Updated M array
}
\description{
Computes posterior mean of M efficiently
}
\keyword{internal}
