% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_static.R
\name{update_theta_tucker}
\alias{update_theta_tucker}
\title{Update Theta Using Tucker Products}
\usage{
update_theta_tucker(Z, M, B, s2)
}
\arguments{
\item{Z}{centered data array (n_row x n_col x p x Tt)}

\item{M}{baseline mean (n_row x n_col x p)}

\item{B}{list of K influence matrices}

\item{s2}{observation variance}
}
\value{
list with updated Y (theta deviations) and X (lagged theta)
}
\description{
updates theta array using B matrices via Tucker products
}
\keyword{internal}
