% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_core.R
\name{update_Z_mu}
\alias{update_Z_mu}
\title{Generic Update for Z and Mu}
\usage{
update_Z_mu(pre, g2, FAM, Aarray = NULL, Barray = NULL)
}
\arguments{
\item{pre}{Current preprocessing object}

\item{g2}{Variance for mu}

\item{FAM}{Family object}

\item{Aarray}{Optional A array for binary family}

\item{Barray}{Optional B array for binary family}
}
\value{
Updated preprocessing object
}
\description{
Updates latent Z values and baseline mu
}
\keyword{internal}
