% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_covariates_accessors.R
\name{vcov.dbn_covariates_fit}
\alias{vcov.dbn_covariates_fit}
\title{Variance-covariance matrix of the covariate coefficient vector}
\usage{
\method{vcov}{dbn_covariates_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dbn_covariates_fit} object.}

\item{...}{Unused.}
}
\value{
A K x K covariance matrix.
}
\description{
Returns the empirical posterior covariance of the beta
draws stored in \code{x$beta}. The result has the standard
\code{vcov} layout: a K x K symmetric matrix.
}
\author{
Tosin Salau and Shahryar Minhas
}
