% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ranklik.R
\name{zscores}
\alias{zscores}
\title{Z-scores from Ranks}
\usage{
zscores(y, ties.method = "average")
}
\arguments{
\item{y}{Ordinal data matrix}

\item{ties.method}{Method for handling ties}
}
\value{
Matrix of z-scores
}
\description{
Computes normal z-scores from ordinal ranks
}
\keyword{internal}
