% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{compare_dbn}
\alias{compare_dbn}
\title{Compare Multiple DBN Models}
\usage{
compare_dbn(...)
}
\arguments{
\item{...}{Multiple dbn objects to compare}
}
\value{
A ggplot2 object or list of plots
}
\description{
Creates comparative plots for multiple DBN results using ggplot2
}
