% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_z_update.R
\name{compute_gaussian_obs_residuals}
\alias{compute_gaussian_obs_residuals}
\title{Compute observation residuals for gaussian family}
\usage{
compute_gaussian_obs_residuals(Z, Theta, M)
}
\arguments{
\item{Z}{observation array (for gaussian, Z = Y)}

\item{Theta}{latent mean array}

\item{M}{baseline mean array}
}
\value{
sum of squared residuals
}
\description{
efficiently computes sum of squared residuals for sigma2_obs update
}
\keyword{internal}
