% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_z_update.R
\name{compute_gaussian_obs_residuals_flat}
\alias{compute_gaussian_obs_residuals_flat}
\title{Vectorized gaussian residual computation}
\usage{
compute_gaussian_obs_residuals_flat(Z_flat, Theta_flat, M_flat, m, p, Tt)
}
\arguments{
\item{Z_flat}{flattened observation array}

\item{Theta_flat}{flattened theta array}

\item{M_flat}{flattened M array}

\item{m}{number of nodes}

\item{p}{number of relations}

\item{Tt}{number of time points}
}
\value{
sum of squared residuals
}
\description{
vectorized version for flattened arrays
}
\keyword{internal}
