% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opt.R
\name{compute_rss_fast}
\alias{compute_rss_fast}
\title{Compute Residual Sum of Squares}
\usage{
compute_rss_fast(Z, M)
}
\arguments{
\item{Z}{Latent values array (m x m x p x Tt)}

\item{M}{Mean array (m x m x p)}
}
\value{
Residual sum of squares
}
\description{
Efficient RSS computation for variance updates
}
\keyword{internal}
