% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_theta_static}
\alias{compute_theta_static}
\title{Compute theta for static model}
\usage{
compute_theta_static(B, Z, M)
}
\arguments{
\item{B}{Sender/receiver effect matrix (m x m)}

\item{Z}{Latent positions array (m x m x p x Tt) or specific slice}

\item{M}{Baseline mean array (m x m x p) or specific slice}
}
\value{
Theta array with same dimensions as Z
}
\description{
Compute theta values on-demand for static DBN model
}
\keyword{internal}
