% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_blocks.R
\name{compute_waic_dbn}
\alias{compute_waic_dbn}
\title{Compute WAIC for DBN Fit}
\usage{
compute_waic_dbn(fit)
}
\arguments{
\item{fit}{dbn model fit object}
}
\value{
list with waic, lppd, p_waic
}
\description{
computes Watanabe-Akaike Information Criterion
}
\keyword{internal}
