% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_structure.R
\name{dag_structure}
\alias{dag_structure}
\title{Convert a formula into an adjacency matrix}
\usage{
dag_structure(fm, coll = NULL)
}
\arguments{
\item{fm}{A formula object with no left hand side. See Details in 
\code{\link{dbn}}.}

\item{coll}{An optional \code{assertCollection} object for returning
validation errors.}
}
\value{
When successful, a list is returned with the following elements
\itemize{
 \item{\code{edge} }{A two column matrix. The first column lists the 
   nodes in the network. The second column shows the parents for the
   respective nodes. The parents are separated by a \code{*}}
 \item{\code{raw_edge} }{Similar to \code{edge}, but the nodes and parents
   also display the time dependency structure.}
 \item{\code{root_node} }{A vector naming the nodes in the network.}
 \item{\code{raw_node} }{Similar to \code{node}, but the nodes are 
   displayed with the \code{[t]} suffix if they are dynamic.}
 \item{\code{root_parent} }{A list of vectors, one for each node, naming the 
   parents of the node.}
 \item{\code{raw_parent} }{A list of vectors, one for each node, naming the
   parents of the node, but with the time dependency suffix.}
 \item{\code{is_dynamic} }{A logical vector indicating if each node is 
   dynamic.}
}
}
\description{
Builds the core structure of the \code{dbn} object.
}
\section{Functional Requirements}{

\enumerate{
  \item Return an error when \code{fm} is not a formula.
  \item Return an error when \code{fm} has a left hand side.
  \item Return an error if a dynamic dependency refers to a 
    future time point (such as \code{t + 1}).
  \item Return a list when successful.
}
}

