% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn.R
\name{dbn}
\alias{dbn}
\alias{dbn.formula}
\alias{dbn.list}
\title{Dynamic Belief Network}
\usage{
dbn(nodes, max_t = 0, ...)

\method{dbn}{formula}(nodes, max_t = 0, ...)

\method{dbn}{list}(nodes, max_t = 0, ...)
}
\arguments{
\item{nodes}{Either a \code{formula} or a \code{list} of models. See 
Details.}

\item{max_t}{\code{numeric(1)}, integerish. Determines the maximum 
time points to extend the model. The initial time point is assumed to
be \code{t = 0}.}

\item{...}{Additional arguments to be passed to other methods. 
Current ignored.}
}
\value{
Returns an object of class \code{dbn}.  The object is a list with two 
elements:

\itemize{
  \item{\code{network} } A \code{formula} object that gives the formula 
        used for the network definition.
  \item{\code{node_attr}} A \code{tbl_df} that contains the 
        characteristics of each node in the network.
}

The \code{node_attr} element has the columns:
\itemize{
  \item{\code{node_name} } The reference name for the node, stored as a character.
  \item{\code{parent} } A list of character strings naming each of the 
        parents of the node.
  \item{\code{is_dynamic} } Logical value.
  \item{\code{max_t} } The maximum time point to which dynamic nodes are 
        extended.
  \item{\code{is_decision} } Logical value.
  \item{\code{is_utility} } Logical value.
  \item{\code{is_deterministic} } Logical value.
  \item{\code{model} } The model object for the node.
  \item{\code{node_name_raw} } The full character name of the node.  For 
        static nodes, this is identical to \code{node_name}. For dynamic 
        nodes, this includes dynamic suffixes, such as \code{"x[t]"}.
        This is provided for convenience of reference in internal operations.
  \item{\code{parent_raw} } A list of character strings giving the full name
        for each of the parents of the node. For  
        static nodes, this is identical to \code{parent}. For dynamic 
        nodes, this includes dynamic suffixes, such as \code{"x[t - 1]"}.
}
}
\description{
Construct a Dynamic Belief Network.  DBNs contain information 
  regarding how nodes and variables relate to each other. Relationships 
  may be defined over time with dynamic variables, and may also contain 
  static variables that do not change over time.
}
\details{
The formula object may not have a left hand side.  The right
hand side is built with a syntax where a node is identified on the 
left of a \code{|} (pipe) and its parents on the right.  If a node has 
multiple parents, they are separated by a \code{*} (asterisk).  Thus
\code{q | a * b} identifies the node \code{q} with parents. Add additional
 nodes with the \code{+} operator (e.g. \code{q | a * b + a | c * d}). 
 There is no limit to how often a node may appear in the parents.  Any nodes
 that are listed in the parents but not as a specific node will still be
 identified as a node, but with no parents.
}
\section{Functional Requirements (General)}{

\enumerate{
 \item{A generic that has \code{formula} and \code{list} methods}
 \item{\code{dbn} objects consist of, at a minimum, a network formula
       and a \code{tbl_df} object providing 
       node attributes.}
 \item{The \code{tbl_df} object contains, at a minimum, columns for 
       the node names; the parents; flags for dynamic, decision, utility, and 
       deterministic nodes; and a model object.}
 \item{Provides an argument that accepts an integerish value indicating 
       the number of time periods over which the network may be observed.}
}
}

\section{Functional Requirements (\code{formula} method)}{

\enumerate{
 \item{Dynamic relationships may be defined relative to time \code{t} 
       using syntax such as \code{t - 1}.}
 \item{Dynamic relationships may not be defined to depend on future
       events, i.e. \code{t + 1}.}
 \item{Dynamic relationship time designations must be integerish with
       where \code{t} is non-negative.}
 \item{\code{dbn} returns an object of class \code{dbn}.}
}


\enumerate{
 \item{The list method converts existing model objects into a 
       \code{dbn} structure.}
}
}

