% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn.R
\name{dbn}
\alias{dbn}
\alias{dbn.formula}
\alias{dbn.list}
\title{Dynamic Bayesian Network}
\usage{
dbn(nodes, ...)

\method{dbn}{formula}(nodes, ...)

\method{dbn}{list}(nodes, ...)
}
\arguments{
\item{nodes}{Either a formula that defines the network (see Details) 
or a list of model objects.}

\item{...}{additional arguments for other methods. currently ignored.}
}
\value{
Returns a \code{dbn} object with attributes to be described:
}
\description{
Generate a Dynamic Bayesian Network.  This section needs more 
  detail as well.  As the crux of the package, perhaps a brief introduction
  to the topic and its application.  Maybe wait until the first 
  vignette is written, and copy content from there?
}
\details{
The formula is used to identify nodes and their parentage.  A node 
definition may be of the format \code{~ [node] | [parent1] * [parent2]}. 
Mutliple nodes may be defined by adding them with the \code{+} operator 
(\code{~ node1 | parent1 * parent2 + node2 | parent3}).

It is permissible to include a node only as a parent; it will still be 
identified as a distinct node in the network.  Each node that has at least 
one parent will need to have its own entry in the formula.
}
\section{Functional Requirements}{

\enumerate{
  \item Accept a one sided formula for the network specification or a 
    list of model objects.
  \item Return an object of class \code{dbn}
  \item The \code{dbn} object has as elements the original \code{nodes}
    formula; The adjacency matrix for the network; a table of each node's
    attributes.
  \item The table of node attributes includes the following columns:
    \code{node_name} (character(1)), \code{parents} (character vector),
    \code{is_decision} (logical), \code{is_utility} (logical),
    \code{is_temporal} (logical), \code{model} (an object to assist in 
    simulation of posterior)
}
}
\author{
Benjamin Nutter
}

