% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_hmm.R
\name{dbn_hmm}
\alias{dbn_hmm}
\title{DBN HMM Model}
\usage{
dbn_hmm(
  Y,
  family = c("ordinal", "gaussian", "binary"),
  R = 3,
  nscan = 5000,
  burn = 1000,
  odens = 1,
  delta = rep(1, R),
  seed = 6886,
  verbose = 100,
  time_thin = 1,
  previous = NULL,
  init = NULL,
  ...
)
}
\arguments{
\item{Y}{Data array (nodes x nodes x relations x time)}

\item{family}{Data family (ordinal, gaussian, or binary)}

\item{R}{Number of regimes}

\item{nscan}{Number of iterations of the Markov chain (beyond burn-in)}

\item{burn}{Burn-in for the Markov chain}

\item{odens}{Output density for the Markov chain}

\item{delta}{Dirichlet prior for transition matrix}

\item{seed}{Random seed}

\item{verbose}{Print progress}

\item{time_thin}{Save every nth time point, when 1, save all time points}

\item{previous}{Previous dbn_hmm results to continue from (optional)}

\item{init}{List of initial values: S, A_list, B_list, Pi, sigma2_proc, tau_A2, tau_B2, g2, pi0 (optional)}

\item{...}{Additional arguments}
}
\value{
List containing MCMC results
}
\description{
Regime-switching DBN for large-scale networks (100+ actors, 50+ time steps)
}
