% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_lowrank.R
\name{dbn_lowrank_accurate}
\alias{dbn_lowrank_accurate}
\title{DBN Low-rank Model (More accurate version ... but slow as heck)}
\usage{
dbn_lowrank_accurate(
  Y,
  family = c("ordinal", "gaussian", "binary"),
  r = 2,
  nscan = 5000,
  burn = 1000,
  odens = 1,
  ar1_alpha = TRUE,
  update_rho_alpha = FALSE,
  ar1_B = FALSE,
  update_rho_B = FALSE,
  seed = 6886,
  verbose = 100,
  time_thin = 1,
  previous = NULL,
  init = NULL,
  ...
)
}
\arguments{
\item{Y}{Data array (nodes x nodes x relations x time)}

\item{family}{Data family}

\item{r}{Rank for low-rank factorization}

\item{nscan}{Number of iterations of the Markov chain (beyond burn-in)}

\item{burn}{Burn-in for the Markov chain}

\item{odens}{Output density for the Markov chain}

\item{ar1_alpha}{Use AR(1) for alpha dynamics}

\item{update_rho_alpha}{Update AR coefficient for alpha}

\item{ar1_B}{Use AR(1) for B dynamics}

\item{update_rho_B}{Update AR coefficient for B}

\item{seed}{Random seed}

\item{verbose}{Print progress}

\item{time_thin}{Save every nth time point}

\item{previous}{Previous results to continue from}

\item{init}{Initial values}

\item{...}{Additional arguments (currently unused)}
}
\value{
List containing MCMC results
}
\description{
Fits DBN with low-rank sender effects
}
