% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{dbn_make_family}
\alias{dbn_make_family}
\title{Internal constructor for likelihood "family" objects}
\usage{
dbn_make_family(
  name,
  draw_latent,
  ffbs_wrapper,
  loglik,
  linkinv,
  rgen_obs,
  init_pars = list()
)
}
\description{
Internal constructor for likelihood "family" objects
}
\keyword{internal}
