% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_plot_default.R
\name{dbn_plot_default}
\alias{dbn_plot_default}
\title{Show and Modify \code{dbn} Default Settings for Plots}
\usage{
dbn_plot_default()
}
\description{
Display the current defaults used for plotting a dynamic 
  belief network.  Modifications may also be made.
}
\section{Functional Requirements}{

\enumerate{
  \item The non-assignment method returns a printed string for each node
        type showing the plotting parameters.
}
}

