% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{debug_irf}
\alias{debug_irf}
\title{Debug method for irf function}
\usage{
debug_irf(fit, draw_idx = 1, shock_i = 8, shock_j = 12)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{draw_idx}{Index of posterior draw to debug (default: 1)}

\item{shock_i}{Source node index for shock (default: 8)}

\item{shock_j}{Target node index for shock (default: 12)}
}
\description{
Debug method for irf function
}
