% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{drop_diag}
\alias{drop_diag}
\title{Remove diagonal from sparse matrix}
\usage{
drop_diag(mat)
}
\arguments{
\item{mat}{Sparse or regular matrix}
}
\value{
Matrix with diagonal elements removed (set to 0)
}
\description{
Efficiently remove diagonal elements from a sparse matrix
}
\keyword{internal}
