% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family_ordinal}
\alias{family_ordinal}
\title{Ordinal Family}
\usage{
family_ordinal()
}
\value{
A dbn_family object
}
\description{
Family object for ordinal outcomes using rank likelihood
}
\keyword{internal}
