% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_dbn_threads}
\alias{get_dbn_threads}
\title{Get the current number of threads used by dbn}
\usage{
get_dbn_threads()
}
\value{
Integer number of threads currently set for parallel computation
}
\description{
Returns the number of OpenMP threads currently configured for
parallel MCMC computation. Defaults to 1 (single-threaded).
}
\examples{
get_dbn_threads()
}
\seealso{
\code{\link{set_dbn_threads}}
}
\author{
Tosin Salau and Shahryar Minhas
}
