% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_dbn_threads}
\alias{get_dbn_threads}
\title{Get the current number of threads used by dbn}
\usage{
get_dbn_threads()
}
\value{
Integer number of threads currently set for parallel computation
}
\description{
Get the current number of threads used by dbn
}
\examples{
# Get current thread count
get_dbn_threads()
}
