% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_extract.R
\name{grep_extract}
\alias{grep_extract}
\title{Extract a Character String Using a Regular Expression}
\usage{
grep_extract(x, pattern, invert = FALSE, ...)
}
\arguments{
\item{x}{Character vector.}

\item{pattern}{\code{character(1)}, a regular expression to apply to 
\code{x}}

\item{invert}{\code{logical(1)}, if \code{TRUE}, extract the non-matched 
substrings.}

\item{...}{Additional arguments to pass to \code{\link{gregexpr}}.}
}
\description{
Extracts a substring from a character string based on a 
  regular expression.  This is similar to \code{stringr::str_extract}, 
  but relies only on base \code{R} methods.
}
\section{Functional Requirements}{

\enumerate{
  \item Returns the substring that matches the regular expression.
  \item If no match is generated, return \code{NA}
}
}
\author{
Benjamin Nutter
}

