% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{initialize_hmm_states}
\alias{initialize_hmm_states}
\title{Initialize HMM states}
\usage{
initialize_hmm_states(Tt, K, init_probs = NULL, trans_probs = NULL)
}
\arguments{
\item{Tt}{Number of time points}

\item{K}{Number of states}

\item{init_probs}{Initial state probabilities}

\item{trans_probs}{Transition probability matrix}
}
\value{
Vector of state assignments
}
\description{
Initialize hidden states for HMM model
}
\keyword{internal}
