% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{kron_mult}
\alias{kron_mult}
\title{Kronecker Product with Sparse Option}
\usage{
kron_mult(A, B, use_sparse = FALSE)
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}

\item{use_sparse}{Use sparse matrices for large dimensions}
}
\value{
Kronecker product A ⊗ B
}
\description{
Computes Kronecker product with optional sparse matrices
}
\keyword{internal}
