% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_extract.R
\name{param_summary}
\alias{param_summary}
\title{Summarize scalar parameters}
\usage{
param_summary(fit, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{fit}{A dbn model fit object}

\item{probs}{Probability levels for quantiles}
}
\value{
Data frame with parameter summaries
}
\description{
Compute quantiles for scalar parameter traces
}
\examples{
\donttest{
sim <- simulate_dynamic_dbn(n = 6, time = 5, seed = 1)
fit <- dbn(sim$Y, model = "dynamic", nscan = 200, burn = 100, verbose = FALSE)
ps <- param_summary(fit)
}
}
\seealso{
\code{\link{theta_summary}}, \code{\link{plot_trace}}, \code{\link{derive_draws}}
}
