% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dbn.R
\name{plot.dbn}
\alias{plot.dbn}
\title{Plot Dynamic Bayesian Network Objects}
\usage{
\method{plot}{dbn}(x, custom_node = character(0),
  custom_edge = character(0), custom_general = character(0), ...)
}
\arguments{
\item{x}{A \code{dbn} object.}

\item{custom_node}{A named \code{vector} of customizations to assign to nodes.}

\item{custom_edge}{A named \code{vector} of customizations to assign to edges.}

\item{custom_general}{A \code{vector} of customizations to assign to the 
graph.}

\item{...}{Additional arguments to pass to other methods}
}
\description{
Prepares a plot of a Dynamic Bayesian Network Object using
  the \code{DiagrammeR} package.
}
\section{Functional Requirements}{

\enumerate{
 \item display passage of time over dynamic variables horizontally across the plot
 \item display dependencies vertically on the plot
 \item display static variables above the dynamic variables
}
}

