% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{plot.dbn_irf}
\alias{plot.dbn_irf}
\title{Plot impulse response functions}
\usage{
\method{plot}{dbn_irf}(x, ci_level = 0.95, title = NULL, ...)
}
\arguments{
\item{x}{A dbn_irf object from compute_irf}

\item{ci_level}{Credible interval level (default: 0.95)}

\item{title}{Plot title (default: auto-generated)}

\item{...}{Additional arguments (ignored)}
}
\value{
A ggplot2 object
}
\description{
Creates a plot of IRF with credible intervals
}
