% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_methods.R
\name{plot_hmm}
\alias{plot_hmm}
\title{Plot HMM-DBN results}
\usage{
plot_hmm(x)
}
\arguments{
\item{x}{A dbn object with model="hmm"}
}
\value{
Either a list of ggplot objects or a combined plot (if gridExtra available)
}
\description{
Creates diagnostic plots for HMM-DBN model fits
}
\keyword{internal}
