% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{plot_piecewise_stability}
\alias{plot_piecewise_stability}
\title{Plot Block Stability}
\usage{
plot_piecewise_stability(x, ...)
}
\arguments{
\item{x}{piecewise dbn fit}

\item{...}{additional arguments}
}
\description{
shows Frobenius norm of A across blocks
}
\keyword{internal}
