% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_methods.R
\name{plot_piecewise_trace}
\alias{plot_piecewise_trace}
\title{Trace Plot for Piecewise Model}
\usage{
plot_piecewise_trace(x, ...)
}
\arguments{
\item{x}{piecewise dbn fit}

\item{...}{unused}
}
\value{
A \code{ggplot} object.
}
\description{
ggplot panel of variance-parameter MCMC traces.
}
\keyword{internal}
