% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{plot_ppc_density}
\alias{plot_ppc_density}
\title{Posterior predictive density plot}
\usage{
plot_ppc_density(fit, ppd = NULL, rel = 1, time = NULL, Y_obs = NULL)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{ppd}{Posterior predictive samples}

\item{rel}{Relation index}

\item{time}{Time indices}

\item{Y_obs}{Observed data array (required)}
}
\value{
A ggplot object or NULL
}
\description{
Plot density of observed vs replicated data
}
\examples{
\donttest{
sim <- simulate_static_dbn(n = 6, time = 5, seed = 1)
fit <- dbn(sim$Y, model = "static", nscan = 200, burn = 100, verbose = FALSE)
ppd <- posterior_predict_dbn(fit, draws = 5)
if (requireNamespace("ggplot2", quietly = TRUE)) plot_ppc_density(fit, ppd, Y_obs = sim$Y)
}
}
\seealso{
\code{\link{posterior_predict_dbn}}, \code{\link{plot_ppc_ecdf}}, \code{\link{param_summary}}
}
\author{
Tosin Salau and Shahryar Minhas
}
