% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{plot_ppc_density}
\alias{plot_ppc_density}
\title{Posterior predictive density plot}
\usage{
plot_ppc_density(fit, ppd = NULL, rel = 1, time = NULL, Y_obs = NULL)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{ppd}{Posterior predictive samples}

\item{rel}{Relation index}

\item{time}{Time indices}

\item{Y_obs}{Observed data array (required)}
}
\value{
A ggplot object or NULL
}
\description{
Plot density of observed vs replicated data
}
