% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{plot_ppc_ecdf}
\alias{plot_ppc_ecdf}
\title{Posterior predictive ECDF plot}
\usage{
plot_ppc_ecdf(
  fit,
  ppd = NULL,
  ndraws_plot = 20,
  alpha = 0.3,
  rel = 1,
  time = NULL,
  Y_obs = NULL
)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{ppd}{Posterior predictive samples (from posterior_predict_dbn)}

\item{ndraws_plot}{Number of replicates to show (default: 20)}

\item{alpha}{Transparency for replicate lines (default: 0.3)}

\item{rel}{Relation index to plot (default: 1)}

\item{time}{Time index to plot (default: all)}

\item{Y_obs}{Observed data array (required)}
}
\value{
A ggplot object
}
\description{
Plot empirical CDFs of observed vs replicated data
}
