% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plots.R
\name{plot_regime_probs}
\alias{plot_regime_probs}
\title{Plot regime probabilities (HMM only)}
\usage{
plot_regime_probs(fit)
}
\arguments{
\item{fit}{A dbn_hmm model fit object}
}
\value{
A ggplot object or NULL
}
\description{
Plot posterior regime probabilities over time
}
