% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plots.R
\name{plot_theta}
\alias{plot_theta}
\title{Plot Theta Heatmap}
\usage{
plot_theta(fit, time = 1, rel = 1, fun = mean, ...)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{time}{Time point to plot}

\item{rel}{Relation to plot}

\item{fun}{Summary function (default: mean)}

\item{...}{Additional arguments to theta_summary}
}
\value{
A ggplot object or NULL
}
\description{
Posterior mean of Theta as a heatmap
}
\examples{
\donttest{
sim <- simulate_dynamic_dbn(n = 6, time = 5, seed = 1)
fit <- dbn(sim$Y, model = "dynamic", nscan = 200, burn = 100, verbose = FALSE)
if (requireNamespace("ggplot2", quietly = TRUE)) plot_theta(fit)
}
}
\seealso{
\code{\link{theta_summary}}, \code{\link{plot_trace}}, \code{\link{param_summary}}
}
