% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plots.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot parameter trace plots}
\usage{
plot_trace(fit, pars = NULL, ncol = 2)
}
\arguments{
\item{fit}{A dbn model fit object}

\item{pars}{Character vector of parameter names to plot}

\item{ncol}{Number of columns for multi-panel plot}
}
\value{
A ggplot object or NULL
}
\description{
Create trace plots for scalar parameters
}
