% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_methods.R
\name{predict_hmm}
\alias{predict_hmm}
\title{Predict from HMM-DBN Fit}
\usage{
predict_hmm(object, H = 1, draws = 100, summary = c("mean", "none"))
}
\arguments{
\item{object}{A dbn object with model="hmm"}

\item{H}{Number of forecast steps}

\item{draws}{Number of posterior draws}

\item{summary}{"mean" for posterior mean, "none" for all draws}
}
\value{
Prediction array
}
\description{
Generates H-step-ahead forecasts from an HMM-DBN model
}
\keyword{internal}
