% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowrank_methods.R
\name{predict_lowrank}
\alias{predict_lowrank}
\title{Forecast / replicate from a low-rank fit}
\usage{
predict_lowrank(object, H = 1, draws = 100, summary = c("mean", "none"))
}
\arguments{
\item{object}{A dbn object with model="lowrank"}

\item{H}{Number of time steps ahead to forecast}

\item{draws}{Number of posterior draws to use}

\item{summary}{Character string: "mean" for posterior mean, "none" for all draws}
}
\value{
Array of predictions with dimensions appropriate to summary choice
}
\description{
Generate predictions from a low-rank DBN model
}
\keyword{internal}
