% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{print.dbn_irf}
\alias{print.dbn_irf}
\title{Print method for dbn_irf objects}
\usage{
\method{print}{dbn_irf}(x, digits = 3, ...)
}
\arguments{
\item{x}{A dbn_irf object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for dbn_irf objects
}
