% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{print.dbn_ppd}
\alias{print.dbn_ppd}
\title{Print method for posterior predictive distribution}
\usage{
\method{print}{dbn_ppd}(x, ...)
}
\arguments{
\item{x}{An object of class "dbn_ppd"}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for posterior predictive distribution
}
