% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_extract.R
\name{regime_probs}
\alias{regime_probs}
\title{Extract regime probabilities for HMM models}
\usage{
regime_probs(fit)
}
\arguments{
\item{fit}{A dbn_hmm model fit object}
}
\value{
Matrix of regime probabilities (T x R) or NULL
}
\description{
Compute posterior probabilities of regime assignments
}
\examples{
\donttest{
sim <- simulate_hmm_dbn(n = 6, time = 10, R = 2, seed = 1)
fit <- dbn(sim$Y, model = "hmm", R = 2, nscan = 200, burn = 100, verbose = FALSE)
rp <- regime_probs(fit)
}
}
\seealso{
\code{\link{plot_regime_probs}}, \code{\link{param_summary}}, \code{\link{dbn_hmm}}
}
\author{
Tosin Salau and Shahryar Minhas
}
