% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_extract.R
\name{regime_probs}
\alias{regime_probs}
\title{Extract regime probabilities for HMM models}
\usage{
regime_probs(fit)
}
\arguments{
\item{fit}{A dbn_hmm model fit object}
}
\value{
Matrix of regime probabilities (T x R) or NULL
}
\description{
Compute posterior probabilities of regime assignments
}
