% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{role_trajectory}
\alias{role_trajectory}
\title{Track Role Evolution}
\usage{
role_trajectory(fit, mat = c("A", "B"), comp = 1)
}
\arguments{
\item{fit}{Dynamic dbn object}

\item{mat}{"A" or "B"}

\item{comp}{Component index (default: 1)}
}
\value{
Base R plot (invisible NULL)
}
\description{
Tracks first left/right singular vector of A_t or B_t
}
