% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ranklik.R
\name{rz_fc}
\alias{rz_fc}
\title{Sample Z given Rank Constraints}
\usage{
rz_fc(R, Z, EZ, IR)
}
\arguments{
\item{R}{Ordinal ranks}

\item{Z}{Current Z values}

\item{EZ}{Expected Z values}

\item{IR}{List of indices for each rank}
}
\value{
Updated Z matrix
}
\description{
Samples latent Z values given ordinal constraints via truncated normal
}
\keyword{internal}
