% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{set_dbn_threads}
\alias{set_dbn_threads}
\title{Set the number of threads used by dbn}
\usage{
set_dbn_threads(n_threads = NULL)
}
\arguments{
\item{n_threads}{Integer number of threads to use for parallel computation.
Use NULL to reset to default (1 thread).}
}
\value{
The previous value of n_threads (invisibly)
}
\description{
Set the number of threads used by dbn
}
\examples{
set_dbn_threads(4)
set_dbn_threads(parallel::detectCores() / 2)
set_dbn_threads(NULL)
}
