% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_piecewise.R
\name{simulate_piecewise_dbn}
\alias{simulate_piecewise_dbn}
\title{Simulate Data from Piecewise DBN}
\usage{
simulate_piecewise_dbn(
  n = 20,
  time = 50,
  blocks = 4,
  p = 1,
  sigma2 = 1,
  tau2 = 0.5,
  seed = NULL
)
}
\arguments{
\item{n}{number of actors}

\item{time}{number of time points}

\item{blocks}{block specification}

\item{p}{number of relations}

\item{sigma2}{observation variance}

\item{tau2}{A/B prior variance}

\item{seed}{random seed}
}
\value{
list with Y, true_A, true_B, true_M, block_info
}
\description{
generates synthetic data from piecewise-static model
}
\examples{
\donttest{
sim <- simulate_piecewise_dbn(n = 10, time = 40, blocks = c(15, 30, 40))
dim(sim$Y)
}
}
