% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_test_data}
\alias{simulate_test_data}
\title{Simulate Simple Test Data}
\usage{
simulate_test_data(n = 10, n_col = n, p = 2, time = 20, seed = NULL)
}
\arguments{
\item{n}{Number of row actors / senders}

\item{n_col}{Number of column actors / receivers (default: n)}

\item{p}{Number of relations}

\item{time}{Number of time points}

\item{seed}{Random seed}
}
\value{
Array of ordinal network data
}
\description{
Generates simple ordinal network data for quick testing
}
\examples{
Y <- simulate_test_data(n = 10, time = 5, seed = 42)
str(Y)
}
\seealso{
\code{\link{dbn}} for model fitting, \code{\link{simulate_static_dbn}} for full simulation with true parameters
}
\author{
Tosin Salau and Shahryar Minhas
}
