% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_test_data}
\alias{simulate_test_data}
\title{Simulate Simple Test Data}
\usage{
simulate_test_data(n = 10, p = 2, time = 20, seed = NULL)
}
\arguments{
\item{n}{Number of actors}

\item{p}{Number of relations}

\item{time}{Number of time points}

\item{seed}{Random seed}
}
\value{
Array of ordinal network data
}
\description{
Generates simple test data for quick testing
}
