% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{sparse_diag}
\alias{sparse_diag}
\title{Create sparse diagonal matrix}
\usage{
sparse_diag(x, sparse_threshold = 100)
}
\arguments{
\item{x}{Diagonal values or dimension}

\item{sparse_threshold}{Size threshold for using sparse format}
}
\value{
Diagonal matrix
}
\description{
Create diagonal matrix, using sparse format if large
}
\keyword{internal}
