% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_safe.R
\name{sparse_quad}
\alias{sparse_quad}
\title{Sparse-aware quadratic form}
\usage{
sparse_quad(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Matrix (often diagonal/sparse)}
}
\value{
Quadratic form result
}
\description{
Compute t(A) \%\emph{\% B \%}\% A efficiently
}
\keyword{internal}
