% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{stat_density}
\alias{stat_density}
\title{Network statistic: density}
\usage{
stat_density(X)
}
\arguments{
\item{X}{Network matrix}
}
\value{
Scalar density value
}
\description{
Compute network density (mean of off-diagonal elements)
}
