% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{stat_in_degree}
\alias{stat_in_degree}
\title{Network statistic: in-degree}
\usage{
stat_in_degree(X)
}
\arguments{
\item{X}{Network matrix}
}
\value{
Vector of in-degrees
}
\description{
Compute in-degree for all nodes (column sums)
}
