% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{stat_out_degree}
\alias{stat_out_degree}
\title{Network statistic: out-degree}
\usage{
stat_out_degree(X)
}
\arguments{
\item{X}{Network matrix}
}
\value{
Vector of out-degrees
}
\description{
Compute out-degree for all nodes (row sums)
}
