% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{stat_reciprocity}
\alias{stat_reciprocity}
\title{Network statistic: reciprocity}
\usage{
stat_reciprocity(X)
}
\arguments{
\item{X}{Network matrix}
}
\value{
Scalar reciprocity value
}
\description{
Compute network reciprocity (correlation between X[i,j] and X[j,i])
}
