% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{stat_transitivity}
\alias{stat_transitivity}
\title{Network statistic: transitivity}
\usage{
stat_transitivity(X)
}
\arguments{
\item{X}{Network matrix}
}
\value{
Scalar transitivity value
}
\description{
Compute transitivity (clustering coefficient)
}
