% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_methods.R
\name{tidy_dbn}
\alias{tidy_dbn}
\title{Tidy DBN Summary}
\usage{
tidy_dbn(fit, what = c("A", "B", "Theta"), time_subset = NULL)
}
\arguments{
\item{fit}{DBN object}

\item{what}{Components to extract}

\item{time_subset}{Time points to include (dynamic model)}
}
\value{
List of posterior mean arrays
}
\description{
Extract posterior means in tidy format
}
\keyword{internal}
