% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opt.R
\name{update_AB_optimized}
\alias{update_AB_optimized}
\title{Update A and B with Pre-allocated Memory}
\usage{
update_AB_optimized(
  Theta_all,
  Aarray,
  Barray,
  sigma2,
  tauA2,
  tauB2,
  ar1 = FALSE,
  rhoA = 0,
  rhoB = 0
)
}
\arguments{
\item{Theta_all}{Full Theta array (m x m x p x Tt)}

\item{Aarray}{Current A array (m x m x Tt)}

\item{Barray}{Current B array (m x m x Tt)}

\item{sigma2}{Observation variance}

\item{tauA2}{A innovation variance}

\item{tauB2}{B innovation variance}

\item{ar1}{Use AR(1) dynamics}

\item{rhoA}{AR(1) coefficient for A}

\item{rhoB}{AR(1) coefficient for B}
}
\value{
List with updated A and B arrays
}
\description{
A/B updates using batch C++ implementation
}
\keyword{internal}
